updated consumer api request and moved all types to types folder
