+++ frontend/src/api/campaigns.ts
@@ -1,63 +1,9 @@
-import { apiUrl } from './config';
-import { getToken } from './auth';
-
-// ---------- Types ----------
-
-export type CampaignStatus = 'draft' | 'active' | 'paused' | 'completed';
-
-/** Shape returned by GET /campaigns and GET /campaigns/:id */
-export interface Campaign {
-  id: number;
-  business_client_id: number;
-  name: string;
-  status: CampaignStatus;
-  budget_total: string | null;
-  start_date: string | null;
-  end_date: string | null;
-  goal: string | null;
-  target_audience: string | null;
-  product_context: string | null;
-  created_at: string;
-  updated_at: string;
-  product_ids: string | null;
-}
-
-/** Payload for POST /campaigns */
-export interface CreateCampaignPayload {
-  business_client_id: number;
-  name: string;
-  status?: CampaignStatus;
-  budget_total?: string | null;
-  start_date?: string | null;
-  end_date?: string | null;
-  goal?: string | null;
-  target_audience?: string | null;
-  product_context?: string | null;
-  product_ids?: string | null;
-}
-
-/** Payload for PUT /campaigns/:id */
-export interface UpdateCampaignPayload {
-  name?: string;
-  status?: CampaignStatus;
-  budget_total?: string | null;
-  start_date?: string | null;
-  end_date?: string | null;
-  goal?: string | null;
-  target_audience?: string | null;
-  product_context?: string | null;
-  product_ids?: string | null;
-}
-
-// ---------- Auth header ----------
-
-function authHeaders(): HeadersInit {
-  const token = getToken();
-  return {
-    'Content-Type': 'application/json',
-    ...(token ? { Authorization: `Bearer ${token}` } : {}),
-  };
-}
+import { apiUrl, authHeaders } from './config';
+import type {
+  Campaign,
+  CreateCampaignPayload,
+  UpdateCampaignPayload
+} from '../types';
 
 // ---------- API calls ----------
