+++ frontend/src/types/health.ts
@@ -1,0 +1,6 @@
+/**
+ * Health check response
+ */
+export interface HealthResponse {
+    status: string;
+}
